requirements.txt
scapy==2.4.3
requests==2.24.0
scapy==2.4.3
urllib3==1.25.9
python-nmap==0.6.1